change to int32s, broken neon implementation

+++ ntt_test.go
@@ -7,14 +7,15 @@ import (
 
 func TestNTTRoundTrip(t *testing.T) {
 	params := ML_DSA_44_Parameters
-	input := make([]int, 256)
+	input := make([]int32, 256)
 	for i := range 256 {
-		input[i] = rand.Intn(params.Q) - params.Q/2
+		input[i] = int32(rand.Intn(int(params.Q))) - params.Q/2
 	}
 	result := nttInverse(params, ntt(params, input))
 
 	for i := range 256 {
 		if result[i] != input[i] {
+			// fmt.Printf("%v", result)
 			t.Fatalf("mismatch at %d: %d != %d", i, result[i], input[i])
 		}
 	}
